Attempt to fix int-long type mismatch in cython_nms

+++ yolact/utils/cython_nms.pyx
@@ -32,8 +32,8 @@ def nms(np.ndarray[np.float32_t, ndim=2] dets, np.float32_t thresh):
     cdef np.ndarray[np.int64_t, ndim=1] order = scores.argsort()[::-1]
 
     cdef int ndets = dets.shape[0]
-    cdef np.ndarray[int, ndim=1] suppressed = \
-            np.zeros((ndets), dtype=int)
+    cdef np.ndarray[np.int64_t, ndim=1] suppressed = \
+            np.zeros((ndets), dtype=np.int64_t)
 
     # nominal indices
     cdef int _i, _j
